% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkins-data.r
\docType{data}
\name{checkins}
\alias{checkins}
\title{An example checkin dataset.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 14609 rows and 3 columns.
}
\usage{
data(checkins)
}
\description{
The checkins data set is an example of a checkin data set.
It consists of three columns.
\enumerate{
\item id: The identifier of a device or person checking into a location.
\item timestamp:The time when the checkin occured.
\item location:The identifier of where a checkin occured.
}

Please note that this data set was simulated illustrative purposes only.
}
\keyword{datasets}
