% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval-iterator.r
\name{hour_checkin_gen}
\alias{hour_checkin_gen}
\title{Create an hour interval iterator generator}
\usage{
hour_checkin_gen(x, time, end = NULL)
}
\arguments{
\item{x}{the data frame of checkins.}

\item{time}{the column denoting the timestamp.}

\item{end}{the end of the interval.}
}
\description{
The hour interval iterator generator is a special case of
the time interval generator, which iterates over consecutive hours.
While this function may be useful on it's own, if you are interested in
applying a function to (possibly grouped) data, then you may be better off
reading the \code{map_hourly_interval} documentation.
}
\examples{

data(checkins)

# Get the first hour entries from the beginning of the checkins data.
hour_checkin_gen(head(checkins), "timestamp")$nextElem()

}
\seealso{
map_hourly_interval
}
