% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkin.r
\name{normalize_by_day}
\alias{normalize_by_day}
\title{Normalize data by day}
\usage{
normalize_by_day(x, ts_col, last_day = NULL)
}
\arguments{
\item{x}{the data frame of checkins}

\item{ts_col}{the column denoting the timestamp}

\item{last_day}{the last day to normalize. If NULL then the last day
in x$timestamp is used.}
}
\description{
Normalize data by day
}
