% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imsi-util.r
\name{checkins_in_interval}
\alias{checkins_in_interval}
\title{Get the checkins in an interval}
\usage{
checkins_in_interval(
  x,
  id_col,
  ts_col,
  start,
  end,
  start_loc = TRUE,
  end_loc = TRUE
)
}
\arguments{
\item{x}{the data frame of checkins}

\item{id_col}{the column name denoting the id}

\item{ts_col}{the column denoting the timestamp}

\item{start}{the beginning of the interval}

\item{end}{the end of the interval}

\item{start_loc}{should the starting location be included? Default TRUE.}

\item{end_loc}{should the starting location be included? Default TRUE.}
}
\description{
Get the checkins in an interval
}
